function that returns the number of bits you would need to flip to get from one number to another

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -7,8 +7,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0;
-	unsigned int a = 0;
+	unsigned long int count = 0;
+	unsigned long int a = 0;
 
 	count = n ^ m;
 	while (count >= 1)
